inventario/requirements.txt
absl-py==2.3.1
alembic==1.18.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asciimatics==1.15.0
asgiref==3.11.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.12.2
bidict==0.23.1
blinker==1.7.0
brotli==1.2.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
diagrams==0.24.1
Django==5.2.11
djangorestframework==3.16.1
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.7
firebase_admin==7.2.0
Flask==3.0.0
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
gevent==25.9.1
geventhttpclient==2.3.9
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.20.3
greenlet==3.2.4
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
locust==2.43.3
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
msgpack==1.1.2
mysqlclient==2.2.8
namex==0.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.3.1
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.3.0
parso==0.8.4
Pillow==10.0.1
platformdirs==4.3.8
plotly==6.5.0
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.33.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyfiglet==1.0.2
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.11.0
PyMsgBox==1.0.9
PyMySQL==1.1.2
pyparsing==3.2.3
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.20
python-socketio==5.16.1
pytweening==1.0.7
pytz==2025.2
pywhatkit==5.4
pywin32==306
pyzmq==27.0.0
referencing==0.37.0
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.28.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.44
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.45.3
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.1.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.12
websocket-client==1.9.0
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
